\name{law}
\alias{law}

\title{
A dataset for test.
}
\description{
The dataset includes two variables (LSAT, GPA) of length 15.
}
\usage{
data(law)
}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{\code{LSAT}}{LSAT (numeric)}
    \item{\code{GPA}}{GPA (numeric)}
  }
}

\keyword{datasets}

