% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCpbeta.R
\name{pbeta.MC}
\alias{pbeta.MC}
\title{A Monte Carlo estimate of the Beta distribution cdf}
\usage{
pbeta.MC(x, p, q, m = 10000)
}
\arguments{
\item{x}{Value of quantiles.}

\item{p, q}{Non-negative integer parameters of the Beta distribution.}

\item{m}{Sampling times}
}
\value{
A Monte Carlo estimate.
}
\description{
A Monte Carlo estimate of the Beta distribution cdf.
}
\examples{
\dontrun{
x<-numeric(9)
for (i in 1:9) x[i]<-pbeta.MC(i/10,3,3)
plot(1:9/10,x,type='l')
}
}
