% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHrw.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{A Metropolis-Hastings sampler from a standard Cauchy distribution using R}
\usage{
rw.Metropolis(sigma = 2, x0 = 10, N)
}
\arguments{
\item{sigma}{The standard deviation of Metropolis-Hastings sampler.}

\item{x0}{The initial value of Metropolis-Hastings sampler.}

\item{N}{The number of samples.}
}
\value{
A random sample of size \code{n}.
}
\description{
A Metropolis-Hastings sampler from a standard Cauchy distribution using R.
}
\examples{
\dontrun{
rnR=rw.Metropolis(2,10,2000)
plot(rnR$x,type='l')
}
}
