% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JKse.R
\name{se.Jackknife}
\alias{se.Jackknife}
\title{A Jackknife estimate of the standard error}
\usage{
se.Jackknife(x, y)
}
\arguments{
\item{x, y}{Numeric vectors or data frames.}
}
\value{
A Jackknife estimate of the standard error.
}
\description{
A Jackknife estimate of the standard error of the correlation statistic.
}
\examples{
\dontrun{
data(law)
attach(law)
x <- law$LSAT
y <- law$GPA
se.Jackknife(x,y)
}
}
